change home page contents to search page
