+++ frontend/fingertips-frontend/app/results/search-result-data.ts
