+++ frontend/fingertips-frontend/components/forms/SearchForm/searchActions.ts
@@ -41,5 +41,5 @@ export async function searchIndicator(
   const { indicator } = validatedFields.data;
 
   const searchState = new SearchStateManager({ indicator });
-  redirect(searchState.generatePath('/search/results'), RedirectType.push);
+  redirect(searchState.generatePath('/results'), RedirectType.push);
 }
