+++ frontend/fingertips-frontend/components/pages/results/searchResultsActions.ts
